Resolve test failures
